pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "wheel>=0.43.0"]
build-backend = "setuptools.build_meta"

[project]
name = "intelligent-fraud-detection-chatbot"
version = "0.2.0"
description = "AI-powered insurance fraud detection chatbot integrating FastAPI, LangChain, and AWS (Lambda, SageMaker, RDS, S3)."
readme = "README.md"
authors = [
  { name = "Pruthviraj Rathod", email = "prithvirathod29884@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Topic :: Security",
  "Topic :: Software Development :: Artificial Intelligence",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.10",
  "Operating System :: OS Independent"
]

dependencies = [
  # --- Core Web ---
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.1",
  "python-multipart>=0.0.9",
  "pydantic>=2.9.2",
  "python-dotenv>=1.0.1",

  # --- ML / AI ---
  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "scikit-learn>=1.5.2",
  "xgboost>=2.1.1",
  "joblib>=1.4.2",

  # --- NLP ---
  "spacy>=3.7.4",
  "sentence-transformers>=3.2.1",
  "textblob>=0.17.1",
  "nltk>=3.9.1",
  "tiktoken>=0.8.0",

  # --- AWS Integration ---
  "boto3>=1.35.22,<1.36.0",
  "awscli>=1.36.12",
  "sagemaker==2.229.0",

  # --- DB / Cache ---
  "SQLAlchemy>=2.0.35",
  "psycopg2-binary>=2.9.9",
  "redis>=5.0.8",

  # --- Security ---
  "PyJWT>=2.9.0",
  "passlib[bcrypt]>=1.7.4",

  # --- Utilities ---
  "requests>=2.32.3",
  "regex>=2024.7.24",
  "geopy>=2.4.1",
  "tqdm>=4.66.5",
  "pytz>=2024.1",

  # --- Visualization / Frontend ---
  "matplotlib>=3.9.2",
  "seaborn>=0.13.2",
  "streamlit>=1.28.1",
  "streamlit-chat>=0.1.1",
  "plotly>=5.17.0",
  "pillow>=10.4.0",

  # --- GenAI / LangChain ---
  "openai>=1.109.1",
  "langchain>=0.3.7",
  "langchain-core>=0.3.17",
  "langchain-openai>=0.2.6",
  "langchain-community>=0.3.0",
  "pinecone-client>=4.1.0",

  # --- Testing / Dev ---
  "pytest>=8.3.3",
  "pytest-asyncio>=0.23.7",
  "pytest-mock>=3.14.0",
  "responses>=0.23.3"
]

[project.optional-dependencies]
dev = [
  "black>=24.8.0",
  "flake8>=7.0.0",
  "coverage>=7.6.0",
  "pre-commit>=3.8.0"
]
ml = [
  "tensorflow>=2.13.0",
  "torch>=2.3.0"
]

[project.urls]
Homepage = "https://github.com/prithvi429/intelligent-fraud-detection-chatbot"
Documentation = "https://github.com/prithvi429/intelligent-fraud-detection-chatbot/wiki"
BugTracker = "https://github.com/prithvi429/intelligent-fraud-detection-chatbot/issues"
Changelog = "https://github.com/prithvi429/intelligent-fraud-detection-chatbot/releases"

[project.scripts]
fraud-api = "src.main:run_api"
fraud-train = "ml.train:main"
fraud-chat = "chatbot.agent:run_agent"

[tool.setuptools.packages.find]
include = ["src*", "chatbot*", "ml*", "tests*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
